requirements.txt
numpy==1.22.4
pandas==1.4.2
nltk==3.7
scikit-learn==1.1.1
scikit-surprise==1.1.1
streamlit==1.9.2
fuzzywuzzy==0.18.0
pickleshare==0.7.5